+++ gala/dynamics/actionangle/analyticactionangle.py
@@ -8,11 +8,33 @@ import numpy as np
 from astropy.constants import G
 import astropy.coordinates as coord
 import astropy.units as u
+from astropy.coordinates.matrix_utilities import rotation_matrix
+from astropy.utils.decorators import deprecated
 
-__all__ = ['isochrone_to_aa', 'harmonic_oscillator_to_aa']
+# Gala
+import gala.dynamics as gd
+from gala.util import GalaDeprecationWarning
+from gala.tests.optional_deps import HAS_TWOBODY
+
+__all__ = ['isochrone_xv_to_aa', 'isochrone_aa_to_xv',
+           'harmonic_oscillator_xv_to_aa']
 
 
-def isochrone_to_aa(w, potential):
+def F(x, y):
+    z = np.zeros_like(x)
+
+    ix = y > np.pi/2.
+    z[ix] = np.pi/2. - np.arctan(np.tan(np.pi/2. - 0.5*y[ix])/x[ix])
+
+    ix = y < -np.pi/2.
+    z[ix] = -np.pi/2. + np.arctan(np.tan(np.pi/2. + 0.5*y[ix])/x[ix])
+
+    ix = (y <= np.pi/2) & (y >= -np.pi/2)
+    z[ix] = np.arctan(x[ix] * np.tan(0.5*y[ix]))
+    return z
+
+
+def isochrone_xv_to_aa(w, potential):
     """
     Transform the input cartesian position and velocity to action-angle
     coordinates in the Isochrone potential. See Section 3.5.2 in
@@ -22,13 +44,6 @@ def isochrone_to_aa(w, potential):
     This transformation is analytic and can be used as a "toy potential"
     in the Sanders & Binney (2014) formalism for computing action-angle
     coordinates in any potential.
-
-    .. note::
-
-        This function is included as a method of the
-        :class:`~gala.potential.IsochronePotential` and it is recommended
-        to call :meth:`~gala.potential.IsochronePotential.phase_space()`
-        instead.
 
     Parameters
     ----------
@@ -40,12 +55,12 @@ def isochrone_to_aa(w, potential):
 
     Returns
     -------
-    actions : :class:`numpy.ndarray`
-        An array of actions computed from the input positions and velocities.
-    angles : :class:`numpy.ndarray`
-        An array of angles computed from the input positions and velocities.
-    freqs : :class:`numpy.ndarray`
-        An array of frequencies computed from the input positions and velocities.
+    actions : :class:`~astropy.units.Quantity`
+        Actions computed from the input positions and velocities.
+    angles : :class:`~astropy.units.Quantity`
+        Angles computed from the input positions and velocities.
+    freqs : :class:`~astropy.units.Quantity`
+        Frequencies computed from the input positions and velocities.
     """
     from gala.potential import Hamiltonian, PotentialBase, IsochronePotential
 
@@ -56,29 +71,39 @@ def isochrone_to_aa(w, potential):
     GM = (G*potential.parameters['m']).decompose(usys).value
     b = potential.parameters['b'].decompose(usys).value
     E = w.energy(Hamiltonian(potential)).decompose(usys).value
-    E = np.squeeze(E)
+    E = np.atleast_1d(E)
 
     if np.any(E > 0.):
         raise ValueError("Unbound particle. (E = {})".format(E))
 
     # convert position, velocity to spherical polar coordinates
     w_sph = w.represent_as(coord.PhysicsSphericalRepresentation)
-    r, phi, theta = map(np.squeeze, [w_sph.r.decompose(usys).value,
-                                     w_sph.phi.radian,
-                                     w_sph.theta.radian])
+    r, phi, theta = map(
+        np.atleast_1d,
+        [
+            w_sph.r.decompose(usys).value,
+            w_sph.phi.radian,
+            w_sph.theta.radian
+        ]
+    )
 
     ang_unit = u.radian/usys['time']
-    vr, phi_dot, theta_dot = map(np.squeeze,
-                                 [w_sph.radial_velocity.decompose(usys).value,
-                                  w_sph.pm_phi.to(ang_unit).value,
-                                  w_sph.pm_theta.to(ang_unit).value])
-    vtheta = r*theta_dot
+    vr, phi_dot, theta_dot = map(
+        np.atleast_1d,
+        [
+            w_sph.radial_velocity.decompose(usys).value,
+            w_sph.pm_phi.to(ang_unit).value,
+            w_sph.pm_theta.to(ang_unit).value
+        ]
+    )
+    vtheta = r * theta_dot
 
     # ----------------------------
     # Compute the actions
     # ----------------------------
 
-    L_vec = np.squeeze(w.angular_momentum().decompose(usys).value)
+    L_vec = [np.atleast_1d(x)
+             for x in w.angular_momentum().decompose(usys).value]
     Lz = L_vec[2]
     L = np.linalg.norm(L_vec, axis=0)
 
@@ -86,7 +111,7 @@ def isochrone_to_aa(w, potential):
     Jr = GM / np.sqrt(-2*E) - 0.5*(L + np.sqrt(L*L + 4*GM*b))
 
     # compute the three action variables
-    actions = np.array([Jr, Lz, L - np.abs(Lz)])  # Jr, Jphi, Jtheta
+    actions = np.array([Jr, Lz, L - np.abs(Lz)]).reshape((3,) + w.shape)
 
     # ----------------------------
     # Angles
@@ -104,26 +129,13 @@ def isochrone_to_aa(w, potential):
     psi = np.arctan2(np.cos(theta), -np.sin(theta)*r*vtheta/L)
     psi[np.abs(vtheta) <= 1e-10] = np.pi/2.  # blows up for small vtheta
 
-    omega_th = 0.5 * (1 + L/np.sqrt(L*L + 4*GM*b))
+    omega_ratio = 0.5 * (1 + L/np.sqrt(L*L + 4*GM*b))
 
     a = np.sqrt((1+e) / (1-e))
     ap = np.sqrt((1 + e + 2*b/c) / (1 - e + 2*b/c))
 
-    def F(x, y):
-        z = np.zeros_like(x)
-
-        ix = y > np.pi/2.
-        z[ix] = np.pi/2. - np.arctan(np.tan(np.pi/2.-0.5*y[ix])/x[ix])
-
-        ix = y < -np.pi/2.
-        z[ix] = -np.pi/2. + np.arctan(np.tan(np.pi/2.+0.5*y[ix])/x[ix])
-
-        ix = (y <= np.pi/2) & (y >= -np.pi/2)
-        z[ix] = np.arctan(x[ix]*np.tan(0.5*y[ix]))
-        return z
-
-    A = omega_th*thetar - F(a, eta) - F(ap, eta)/np.sqrt(1 + 4*GM*b/L/L)
-    thetaz = psi + A
+    A = omega_ratio*thetar - F(a, eta) - F(ap, eta)/np.sqrt(1 + 4*GM*b/L/L)
+    thetat = psi + A
 
     LR = Lz/L
     sinu = (LR/np.sqrt(1.-LR*LR)/np.tan(theta))
@@ -133,8 +145,8 @@ def isochrone_to_aa(w, potential):
     uu[sinu < -1.] = -np.pi/2.
     uu[vtheta > 0.] = np.pi - uu[vtheta > 0.]
 
-    thetap = phi - uu + np.sign(Lz)*thetaz
-    angles = np.array([thetar, thetap, thetaz])
+    thetap = phi - uu + np.sign(Lz)*thetat
+    angles = np.array([thetar, thetap, thetat]).reshape((3,) + w.shape)
     angles = angles % (2*np.pi)
 
     # ----------------------------
@@ -143,160 +155,128 @@ def isochrone_to_aa(w, potential):
     freqs = np.zeros_like(actions)
     omega_r = GM**2 / (Jr + 0.5*(L + np.sqrt(L*L + 4*GM*b)))**3
     freqs[0] = omega_r
-    freqs[1] = np.sign(actions[1]) * omega_th * omega_r
-    freqs[2] = omega_th * omega_r
+    freqs[1] = np.sign(actions[1]) * omega_ratio * omega_r
+    freqs[2] = omega_ratio * omega_r
 
     a_unit = (1 * usys['angular momentum'] / usys['mass']).decompose(usys).unit
     f_unit = (1 * usys['angular speed']).decompose(usys).unit
-    return actions*a_unit, angles*u.radian, freqs*f_unit
+    return actions * a_unit, angles * u.radian, freqs * f_unit
 
 
-def isochrone_to_xv(actions, angles, potential):
+@deprecated(since="v1.5",
+            name="isochrone_to_aa",
+            alternative="isochrone_xv_to_aa",
+            warning_type=GalaDeprecationWarning)
+def isochrone_to_aa(*args, **kwargs):
     """
-    Transform the input actions and angles to ordinary phase space (position
-    and velocity) in cartesian coordinates. See Section 3.5.2 in
+    Deprecated! Use `gala.dynamics.actionangle.isochrone_xv_to_aa` instead.
+    """
+    return isochrone_xv_to_aa(*args, **kwargs)
+
+
+def isochrone_aa_to_xv(actions, angles, potential):
+    """
+    Transform the input actions and angles to cartesian position and velocity
+    coordinates in the Isochrone potential. See Section 3.5.2 in
     Binney & Tremaine (2008), and be aware of the errata entry for
     Eq. 3.225.
 
-    .. note::
-
-        This function is included as a method of the
-        :class:`~gala.potential.IsochronePotential`
-        and it is recommended to call
-        :meth:`~gala.potential.IsochronePotential.action_angle()` instead.
-
     Parameters
     ----------
-    actions : array_like
-        Action variables. Must have shape ``(3, N)`` or ``(3,)``.
-    angles : array_like
-        Angle variables. Must have shape ``(3, N)`` or ``(3,)``.
-        Should be in radians.
-    potential : :class:`gala.potential.IsochronePotential`
+    actions : :class:`~astropy.units.Quantity`
+    angles : :class:`~astropy.units.Quantity`
+    potential : :class:`gala.potential.IsochronePotential`, dict
         An instance of the potential to use for computing the transformation
-        to angle-action coordinates.
+        to angle-action coordinates. Or, a dictionary of parameters used to
+        define an :class:`gala.potential.IsochronePotential` instance.
 
     Returns
     -------
-    x : :class:`numpy.ndarray`
-        An array of cartesian positions computed from the input
-        angles and actions.
-    v : :class:`numpy.ndarray`
-        An array of cartesian velocities computed from the input
-        angles and actions.
+    w : :class:`gala.dynamics.PhaseSpacePosition`
+        The computed positions and velocities.
     """
+    if not HAS_TWOBODY:
+        raise ImportError(
+            "Failed to import twobody: Converting from action-angle "
+            "coordinates to position and velocity in the isochrone potential "
+            "requires a Kepler solver, and thus `twobody` must be installed.")
 
-    raise NotImplementedError("Implementation not supported until working with "
-                              "angle-action variables has a better API.")
+    import twobody as tb
 
-    # actions = atleast_2d(actions, insert_axis=1).copy()
-    # angles = atleast_2d(angles, insert_axis=1).copy()
+    Jr, Jphi, Jth = [np.atleast_1d(x) for x in actions]
+    thr, thphi, thth = [np.atleast_1d(x) for x in angles]
 
-    # usys = potential.units
-    # GM = (G*potential.parameters['m']).decompose(usys).value
-    # b = potential.parameters['b'].decompose(usys).value
+    GM = (G * potential.parameters['m'])
+    b = potential.parameters['b']
 
-    # # actions
-    # Jr = actions[0]
-    # Lz = actions[1]
-    # L = actions[2] + np.abs(Lz)
+    Lz = Jphi
+    L = Jth + np.abs(Lz)
 
-    # # angles
-    # theta_r, theta_phi, theta_theta = angles
+    # Eq.3.225 in B&T 2008
+    sqrt_L2_4GMb = np.sqrt(L**2 + 4 * GM * b)
+    E = -0.5 * (GM / (Jr + 0.5 * (L + sqrt_L2_4GMb)))**2
 
-    # # get longitude of ascending node
-    # theta_1 = theta_phi - np.sign(Lz)*theta_theta
-    # Omega = theta_1
+    # Coordinates orientation crap
+    i = np.arccos(Lz / L)
+    lon_nodes = coord.Angle(thphi - np.sign(Lz) * thth).wrap_at(2*np.pi*u.rad)
+    # TODO: could check that std(i), std(lon_nodes) are small...
 
-    # # Ly = -np.cos(Omega) * np.sqrt(L**2 - Lz**2)
-    # # Lx = np.sqrt(L**2 - Ly**2 - Lz**2)
-    # cosi = Lz/L
-    # sini = np.sqrt(1 - cosi**2)
+    # Auxiliary variables (Eq. 3.240)
+    c = GM / (-2*E) - b
+    e = np.sqrt(1 - L**2 / (GM * c) * (1 + b/c))
 
-    # # Hamiltonian (energy)
-    # H = -2. * GM**2 / (2.*Jr + L + np.sqrt(4.*b*GM + L**2))**2
+    e_eff = e * c / (c + b)
+    eta = tb.eccentric_anomaly_from_mean_anomaly(thr, e_eff)
 
-    # if np.any(H > 0.):
-    #     raise ValueError("Unbound particle. (E = {})".format(H))
+    s = 2 + c/b * (1 - e * np.cos(eta))
+    r = b * np.sqrt((s - 1)**2 - 1)
 
-    # # Eq. 3.240
-    # c = -GM / (2.*H) - b
-    # e = np.sqrt(1 - L*L*(1 + b/c) / GM / c)
+    Omr = GM**2 / (Jr + 0.5 * (L + sqrt_L2_4GMb))**3
+    eta_dot = Omr / (1 - e_eff * np.cos(eta))
+    s_dot = e * c / b * np.sin(eta) * eta_dot
+    vr = b * (s - 1) * s_dot / np.sqrt((s-1)**2 - 1)
+    v_tan = L / r
 
-    # # solve for eta
-    # theta_3 = theta_r
-    # eta_func = lambda x: x - e*c/(b+c)*np.sin(x) - theta_3
-    # eta_func_prime = lambda x: 1 - e*c/(b+c)*np.cos(x)
+    sqrt1 = np.sqrt(1 + e) / np.sqrt(1-e)
+    sqrt2 = np.sqrt(1 + e + 2*b/c) / np.sqrt(1 - e + 2*b/c)
 
-    # # use newton's method to find roots
-    # niter = 100
-    # eta = np.ones_like(theta_3)*np.pi/2.
-    # for i in range(niter):
-    #     eta -= eta_func(eta)/eta_func_prime(eta)
+    with u.set_enabled_equivalencies(u.dimensionless_angles()):
+        terms = (
+            0.5 * (1 + L / sqrt_L2_4GMb) * thr
+            - F(sqrt1, eta)
+            - L / sqrt_L2_4GMb * F(sqrt2, eta)
+        )
+    # psi = angles[2] - terms
+    psi = thth - terms - 3*np.pi/2*u.rad  # WT actual F
 
-    # # TODO: when to do this???
-    # eta -= 2*np.pi
+    xyz_prime = (np.array([
+        r.value * np.cos(psi),
+        r.value * np.sin(psi),
+        np.zeros_like(r.value)]
+    ) * r.unit).to(potential.units['length'])
 
-    # r = c*np.sqrt((1-e*np.cos(eta)) * (1-e*np.cos(eta) + 2*b/c))
-    # vr = np.sqrt(GM/(b+c))*(c*e*np.sin(eta))/r
+    vx = vr * np.cos(psi) - v_tan * np.sin(psi)
+    vy = vr * np.sin(psi) + v_tan * np.cos(psi)
+    vxyz_prime = (np.array([
+        vx.value,
+        vy.to_value(vx.unit),
+        np.zeros_like(r.value)
+    ]) * vx.unit).to(potential.units['velocity'])
 
-    # theta_2 = theta_theta
-    # Omega_23 = 0.5*(1 + L / np.sqrt(L**2 + 4*GM*b))
+    M1 = rotation_matrix(-i, 'y')
+    M2 = rotation_matrix(-lon_nodes, 'z')
+    M3 = rotation_matrix(np.pi/2 * u.rad, 'z')  # WT actual F
+    M = np.einsum('ij,...jk,...kl->...il', M3, M2, M1)
 
-    # a = np.sqrt((1+e) / (1-e))
-    # ap = np.sqrt((1 + e + 2*b/c) / (1 - e + 2*b/c))
+    xyz = np.einsum('...ij,j...->i...', M, xyz_prime)
+    vxyz = np.einsum('...ij,j...->i...', M, vxyz_prime)
 
-    # def F(x, y):
-    #     z = np.zeros_like(x)
+    w = gd.PhaseSpacePosition(pos=xyz, vel=vxyz)
 
-    #     ix = y>np.pi/2.
-    #     z[ix] = np.pi/2. - np.arctan(np.tan(np.pi/2.-0.5*y[ix])/x[ix])
-
-    #     ix = y<-np.pi/2.
-    #     z[ix] = -np.pi/2. + np.arctan(np.tan(np.pi/2.+0.5*y[ix])/x[ix])
-
-    #     ix = (y<=np.pi/2) & (y>=-np.pi/2)
-    #     z[ix] = np.arctan(x[ix]*np.tan(0.5*y[ix]))
-    #     return z
-
-    # theta_2[Lz < 0] -= 2*np.pi
-    # theta_3 -= 2*np.pi
-    # A = Omega_23*theta_3 - F(a, eta) - F(ap, eta)/np.sqrt(1 + 4*GM*b/L/L)
-    # psi = theta_2 - A
-
-    # # theta
-    # theta = np.arccos(np.sin(psi)*sini)
-    # vtheta = L*sini*np.cos(psi)/np.cos(theta)
-    # vtheta = -L*sini*np.cos(psi)/np.sin(theta)/r
-    # vphi = Lz / (r*np.sin(theta))
-
-    # d_phi = vphi / (r*np.sin(theta))
-    # d_theta = vtheta / r
-
-    # # phi
-    # sinu = np.sin(psi)*cosi/np.sin(theta)
-
-    # uu = np.arcsin(sinu)
-    # uu[sinu > 1.] = np.pi/2.
-    # uu[sinu < -1.] = -np.pi/2.
-    # uu[vtheta > 0.] = np.pi - uu[vtheta > 0.]
-
-    # sinu = cosi/sini * np.cos(theta)/np.sin(theta)
-    # phi = (uu + Omega) % (2*np.pi)
-
-    # # We now need to convert from spherical polar coord to cart. coord.
-    # pos = coord.PhysicsSphericalRepresentation(r=r*u.dimensionless_unscaled,
-    #                                            phi=phi*u.rad, theta=theta*u.rad)
-    # pos = pos.represent_as(coord.CartesianRepresentation)
-    # x = pos.xyz.value
-
-    # vel = coord.PhysicsSphericalDifferential(d_phi=d_phi, d_theta=d_theta, d_r=vr)
-    # v = vel.represent_as(coord.CartesianDifferential, base=pos).d_xyz.value
-
-    # return x, v
+    return w.reshape(actions.shape[1:])
 
 
-def harmonic_oscillator_to_aa(w, potential):
+def harmonic_oscillator_xv_to_aa(w, potential):
     """
     Transform the input cartesian position and velocity to action-angle
     coordinates for the Harmonic Oscillator potential.
@@ -305,18 +285,19 @@ def harmonic_oscillator_to_aa(w, potential):
     in the Sanders & Binney (2014) formalism for computing action-angle
     coordinates in any potential.
 
-    .. note::
-
-        This function is included as a method of the
-        :class:`~gala.potential.HarmonicOscillatorPotential`
-        and it is recommended to call
-        :meth:`~gala.potential.HarmonicOscillatorPotential.action_angle()`
-        instead.
-
     Parameters
     ----------
     w : :class:`gala.dynamics.PhaseSpacePosition`, :class:`gala.dynamics.Orbit`
     potential : Potential
+
+    Returns
+    -------
+    actions : :class:`~astropy.units.Quantity`
+        Actions computed from the input positions and velocities.
+    angles : :class:`~astropy.units.Quantity`
+        Angles computed from the input positions and velocities.
+    freqs : :class:`~astropy.units.Quantity`
+        Frequencies computed from the input positions and velocities.
     """
 
     usys = potential.units
@@ -337,7 +318,7 @@ def harmonic_oscillator_to_aa(w, potential):
     except AttributeError:  # not a Quantity
         omega = potential.parameters['omega'].reshape(_new_omega_shape)
 
-    action = (v**2 + (omega*x)**2)/(2.*omega)
+    action = (v**2 + (omega*x)**2) / (2.*omega)
 
     angle = np.arctan(-v / omega / x)
     angle[x == 0] = -np.sign(v[x == 0])*np.pi/2.
@@ -348,9 +329,21 @@ def harmonic_oscillator_to_aa(w, potential):
     if usys is not None and usys:
         a_unit = (1*usys['angular momentum']/usys['mass']).decompose(usys).unit
         f_unit = (1*usys['angular speed']).decompose(usys).unit
-        return action*a_unit, (angle % (2.*np.pi))*u.radian, freq*f_unit
+        return action * a_unit, (angle % (2.*np.pi)) * u.radian, freq * f_unit
     else:
-        return action*u.one, (angle % (2.*np.pi))*u.one, freq*u.one
+        return action * u.one, (angle % (2.*np.pi)) * u.one, freq * u.one
+
+
+@deprecated(since="v1.5",
+            name="harmonic_oscillator_to_aa",
+            alternative="harmonic_oscillator_xv_to_aa",
+            warning_type=GalaDeprecationWarning)
+def harmonic_oscillator_to_aa(*args, **kwargs):
+    """
+    Deprecated! Use `gala.dynamics.actionangle.harmonic_oscillator_xv_to_aa`
+    instead.
+    """
+    return harmonic_oscillator_xv_to_aa(*args, **kwargs)
 
 
 def harmonic_oscillator_to_xv(actions, angles, potential):